requirements.txt
amqp>=2.0.3
autopep8>=1.2.4
celery>=4
coverage>=4.2
dj-database-url>=0.4.1
django==2.0.7
django-bootstrap3>=7.0.1
django-cors-headers>=2.4.0
django-debug-toolbar==1.9
django-phonenumber-field>=1.1.0
django-safedelete==0.3.3
djangorestframework>=3.8.2
djangorestframework-xml>=1.3.0
flake8>=3.0.4
pycodestyle==2.3.1
gunicorn>=19.6.0
ipdb>=0.10.1
ipython>=5.1.0
ipython-genutils>=0.1.0
newrelic>=2.68.0.50
pep8>=1.7.0
python-dateutil>=2.5.3
pytz>=2016.6.1
redis>=2.10.5
django-redis>=4.7.0
requests>=2.11.1
requests-mock>=1.0.0
selenium>=2.53.6
twilio==5.4.0
requests-mock>=1.0.0
django-filter>=2.0.0
psycopg2-binary>=2.6.1
django-session-security
django-axes==4.4.0
beautifulsoup4==4.6.1
-e git+https://github.com/veox/python3-krakenex.git#egg=krakenx
-e git+https://github.com/byrnereese/uphold-sdk-python.git#egg=uphold
isort
freezegun>=0.3.8
suds-jurko>=0.6
drf-extensions>=0.3.1
mock
requests-cache>=0.4.13
xmltodict>=0.10.2
python-bitcoinrpc
urllib3==1.22
pillow
-e git+git://github.com/skinderis/django-loginurl.git#egg=loginurl
social-auth-app-django==2.1.0
coveralls==1.1
django-countries==4.6.1
flower==0.9.2
django-guardian==1.4.9
dredd-hooks==0.2.0
django-fsm==2.6.0
-e git+https://github.com/ericsomdahl/python-bittrex.git#egg=bittrex
-e git+https://github.com/SarunasAzna/poloniex.git#egg=poloniex
cached-property>=1.3.1
eth-utils==0.7.1
web3==3.16.5
pycrypto==2.6.1
django-extensions==1.9.9
django-oauth-toolkit==1.0.0
iptools==0.6.1
numpy==1.14.1
django-otp==0.4.3
qrcode==6.0
git+git://github.com/vvangelovski/django-audit-log.git@0a3b1a03b2ce0b560859cefab00bdd2af6641d59
rlp==0.6.0
django-newsletter==0.7
